Update README.md
The core issues here are: the lack of definition of "identity" and "verifiable". The former is (cannot be) a subject of the technology itself but provide at best levels of representation each of which (i.e. personas, avatars etc.) have various levels of "verifiability" (authentication, certification etc.). -- Without these distinction the setup is not viable.